pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "test_infra"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "flask == 2.3.2",
  "cloudevents == 1.9.0",
  "gunicorn == 22.0.0"
]

[project.optional-dependencies]
signify-secret-rotator = ["google-cloud-secret-manager == 2.20.2"]
slack-message-sender = [
  "slack_bolt == 1.18.0",
  "slack_sdk == 3.21.3",
  "pygithub == 2.1.1",
  "pyyaml == 6.0.1"
]

[project.scripts]
signify-secret-rotator = "signifysecretrotator.signifysecretrotator:setup_app"

[tool.setuptools]
packages = [
  "slackmessagesender",
  "signifysecretrotator",
  "secretmanager",
  "pylogger"
]
package-dir = {"slackmessagesender" = "cmd/cloud-run/slack-message-sender","signifysecretrotator" = "cmd/cloud-run/signifysecretrotator","" = "pkg"}